Find dish using class property
In appStarWarsMeals.py:

Added class property decorator to 'Character' to find 'dish', instead of having 'self.dish'. This way, 'dish' is not generated for a character until a character is selected by user.

Changed 'self.homeworld' to 'self.homeworldName' for clarification.

Added 'self.planet' to 'Character' so it can be called in property 'dish' function.

Rearranged code for readability and did general reformatting.